+++ src/hydra/HydraAdmin.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
-import {AdminGuesserComponent} from './AdminGuesser';
-import {resources} from './__fixtures__';
+import {AdminGuesser} from './HydraAdmin';
+import {resources} from '../__fixtures__';
 
 describe('<AdminGuesser />', () => {
   const renderer = new ShallowRenderer();
@@ -12,7 +12,7 @@ describe('<AdminGuesser />', () => {
   test('renders errors', () => {
     const tree = renderer.render(
       <Provider store={store}>
-        {AdminGuesserComponent({
+        {AdminGuesser({
           error: 'Failed to fetch documentation',
           resources: resources,
           fetching: false,
@@ -25,9 +25,7 @@ describe('<AdminGuesser />', () => {
 
   test('renders loading', () => {
     const tree = renderer.render(
-      <Provider store={store}>
-        {AdminGuesserComponent({fetching: true})}
-      </Provider>,
+      <Provider store={store}>{AdminGuesser({fetching: true})}</Provider>,
     );
 
     expect(tree).toMatchSnapshot();
@@ -36,7 +34,7 @@ describe('<AdminGuesser />', () => {
   test('renders without custom resources', () => {
     const tree = renderer.render(
       <Provider store={store}>
-        {AdminGuesserComponent({resources: resources, fetching: false})}
+        {AdminGuesser({resources: resources, fetching: false})}
       </Provider>,
     );
 
@@ -45,9 +43,7 @@ describe('<AdminGuesser />', () => {
 
   test('renders without data', () => {
     const tree = renderer.render(
-      <Provider store={store}>
-        {AdminGuesserComponent({fetching: false})}
-      </Provider>,
+      <Provider store={store}>{AdminGuesser({fetching: false})}</Provider>,
     );
 
     expect(tree).toMatchSnapshot();
@@ -56,7 +52,7 @@ describe('<AdminGuesser />', () => {
   test('renders errors without data', () => {
     const tree = renderer.render(
       <Provider store={store}>
-        {AdminGuesserComponent({
+        {AdminGuesser({
           error: 'Failed to fetch documentation',
           fetching: false,
         })}
